+++ dst/app/routes/actions.js
@@ -25,17 +25,16 @@ actionsRouter.use(authentication_1.default);
 /**
  * アクション検索
  */
-actionsRouter.get('', permitScopes_1.default(['admin']), (req, __, next) => {
-    req.checkQuery('startDateFrom')
+actionsRouter.get('', permitScopes_1.default(['admin']), ...[
+    check_1.query('startDateFrom')
         .optional()
         .isISO8601()
-        .toDate();
-    req.checkQuery('startDateThrough')
+        .toDate(),
+    check_1.query('startDateThrough')
         .optional()
         .isISO8601()
-        .toDate();
-    next();
-}, validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
+        .toDate()
+], validator_1.default, (req, res, next) => __awaiter(void 0, void 0, void 0, function* () {
     try {
         const actionRepo = new pecorino.repository.Action(mongoose.connection);
         const actions = yield actionRepo.search(Object.assign(Object.assign({}, req.query), {