+++ src/app/middlewares/validator.spec.ts
@@ -3,6 +3,7 @@
  * バリデーションミドルウェアテスト
  */
 import * as assert from 'assert';
+import * as expressValidator from 'express-validator';
 import * as nock from 'nock';
 import * as sinon from 'sinon';
 
@@ -29,15 +30,23 @@ describe('validator', () => {
             isEmpty: () => undefined
         };
         const params = {
-            req: {
-                getValidationResult: async () => validatorResult
-            },
+            req: {},
             res: {},
             next: () => undefined
         };
 
-        sandbox.mock(validatorResult).expects('isEmpty').once().returns(true);
-        sandbox.mock(params).expects('next').once().withExactArgs();
+        sandbox.mock(expressValidator)
+            .expects('validationResult')
+            .once()
+            .returns(validatorResult);
+        sandbox.mock(validatorResult)
+            .expects('isEmpty')
+            .once()
+            .returns(true);
+        sandbox.mock(params)
+            .expects('next')
+            .once()
+            .withExactArgs();
 
         const result = await validator(<any>params.req, <any>params.res, params.next);
         assert.equal(result, undefined);
@@ -50,15 +59,23 @@ describe('validator', () => {
             array: () => [{ param: 'param', msg: 'msg' }]
         };
         const params = {
-            req: {
-                getValidationResult: async () => validatorResult
-            },
+            req: {},
             res: {},
             next: () => undefined
         };
 
-        sandbox.mock(validatorResult).expects('isEmpty').once().returns(false);
-        sandbox.mock(params).expects('next').once().withExactArgs(sinon.match.instanceOf(APIError));
+        sandbox.mock(expressValidator)
+            .expects('validationResult')
+            .once()
+            .returns(validatorResult);
+        sandbox.mock(validatorResult)
+            .expects('isEmpty')
+            .once()
+            .returns(false);
+        sandbox.mock(params)
+            .expects('next')
+            .once()
+            .withExactArgs(sinon.match.instanceOf(APIError));
 
         const result = await validator(<any>params.req, <any>params.res, params.next);
         assert.equal(result, undefined);